% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdev_job.R
\name{new_rdev_job}
\alias{new_rdev_job}
\title{Construct an R development job}
\usage{
new_rdev_job(data, id = "job")
}
\arguments{
\item{data}{A non-empty data frame with at least one numeric column.}

\item{id}{A safe scalar file stem used in user-facing artifact names.}
}
\value{
An object of class \code{rdev_job}.
}
\description{
Construct an R development job
}
