% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdev_job.R
\name{run_job}
\alias{run_job}
\title{Run a development job}
\usage{
run_job(x, ...)
}
\arguments{
\item{x}{An object to execute.}

\item{...}{Arguments passed to a class-specific method.}
}
\value{
The updated job object.
}
\description{
Run a development job
}
