% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdev_job.R
\name{validate_rdev_job}
\alias{validate_rdev_job}
\title{Validate an R development job}
\usage{
validate_rdev_job(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
\code{x}, invisibly, or an error describing the broken invariant.
}
\description{
Validate an R development job
}
